requirements.txt
fastapi~=0.104.1
uvicorn~=0.24.0.post1
starlette~=0.27.0
python-dotenv~=1.0.0
SQLAlchemy~=2.0.15
psycopg2-binary==2.9.9
PyYAML~=6.0.1
requests~=2.31.0
mlflow~=2.7.0
python-multipart=0.0.6
pandas~=2.1.3
pydantic~=1.10.13